Rename .gitlab-ci.yml to db-testing.yml
Since all the jobs in this file relate to running
DB testing, we can create a separate file for this.
The ops mirror will be configured to use this file
when starting pipelines.

We can later add a .gitlab-ci.yml which has jobs like
testing, linting, dependency scanning, etc. that runs
on gitlab.com.