Pretend we are GitLab.com
This is a workaround to also execute migrations that are only guarded by
a `Gitlab.com?` check. Those wouldn't run in the testing pipeline at the
moment.

However, this is considered an anti-pattern in any case. Those guards
should come with the additional `Gitlab.com? || Gitlab.dev_or_test_env?`
guard to make sure those execute on CI, too.

Longer-term we'd like to move away from those guards, see
https://gitlab.com/gitlab-org/gitlab/-/issues/225101#note_643524956.

For now, we'll patch the testing pipeline explicitly which seems the
simplest way of overriding this. For alternative options see
https://gitlab.com/gitlab-org/database-team/gitlab-com-database-testing/-/issues/25#note_656102134.

Closes
https://gitlab.com/gitlab-org/database-team/gitlab-com-database-testing/-/issues/25