Prevent clone id collisions when jobs are retried
When the migration testing job is retried, the pipeline id is the same
as the previous job execution. This causes a failure because clone ids
need to be unique, and the previous clone hasn't been torn down yet.

Include the CI_JOB_ID in the clone id to ensure that it is unique on
every retry.