Add support for multiple clone details
Solves #45

When we test a migration against clones in dblab, we need the pipeline
to annotate the originating MR with a comment providing links to the
clones that were used to perform the testing. Before, we only provided a
working link to the postgres-main clone. This change adds a link to the
postgres-ci clone as well.

Summary of changes:

* Removal of "decomposed" verbiage and variables from pipeline
definitions - we always run "decomposed" now.

* "Collapsing" of db:migrations in db-testing.yml to remove the
"decomposed" extension and merge it back into the parent object.

* Update of gitlab container's clone details script to track the
hostname, project name and instance ID necessary to build a URL for each
clone, and annotate that information to the clone details file.

* Update the db-testing.yml to not pass anything to the clone details
script. We manage this with a yml file on the gitlab container now.

* Update the 'notify' app to treat clone details as an array as opposed
to a single object. Also update spec files and test fixtures to reflect
this.

* Update clone details erb template to loop over array and use new
instanceId and projectName variables to build requisite URLs.