Streamline test patch application
While working issue #45, we discovered that the test patches were
failing to apply, failing silently (the pipeline still returned a green
status), and not specifying which patch file failed.

This indicates that that the files we're patching in gitlab-com have
drifted "out of sync" with this set of patches, and when this happens in
the future we need a way to quickly detect and fix the offending patch.

This commit does the following:

* Renames patches in order of application for easy collated shell
expansion.

* Adds a new script to apply the patches using shell expansion. This
makes adding/removing new patches easier.

* Uses `set -e` in the new script to ensure that failing patches fail
loudly.